chore: update to v1, add device options
