+++ docs/flutter/authenticator/lib/stubs/amplify_auth_cognito_stub.dart
@@ -1,24 +1,14 @@
-import 'dart:async';
 import 'dart:core';
 import 'dart:math';
 import 'package:amplify_auth_cognito/amplify_auth_cognito.dart';
-import 'package:amplify_auth_cognito/amplify_auth_cognito_stream_controller.dart';
 import 'package:amplify_core/amplify_core.dart';
-export 'package:amplify_auth_cognito/src/types.dart';
 
 /// A stub of [AmplifyAuthCognito] that creates users in memory.
 class AmplifyAuthCognitoStub extends AuthPluginInterface
     implements AmplifyPluginInterface {
   AmplifyAuthCognitoStub() : super();
 
-  static AuthStreamController streamWrapper = _AuthStreamControllerStub();
-
   static set instance(AuthPluginInterface instance) {}
-
-  @override
-  StreamController<AuthHubEvent> get streamController {
-    return streamWrapper.authStreamController;
-  }
 
   @override
   Future<void> addPlugin() async {}
@@ -28,7 +18,7 @@ class AmplifyAuthCognitoStub extends AuthPluginInterface
     await Future<void>.delayed(const Duration(milliseconds: 50));
     _MockUser? user = _users[request.username];
     if (user != null) {
-      throw const UsernameExistsException('_MockUser already exists.');
+      throw UsernameExistsException();
     } else {
       _MockUser newUser = _MockUser(
         sub: Random().nextInt(10000).toString(),
@@ -55,9 +45,9 @@ class AmplifyAuthCognitoStub extends AuthPluginInterface
   }) async {
     await Future<void>.delayed(const Duration(milliseconds: 50));
     if (request.confirmationCode != '123456') {
-      throw const CodeMismatchException('Incorrect code. Please try again.');
+      throw CodeMismatchException();
     }
-    return CognitoSignUpResult(
+    return const CognitoSignUpResult(
       isSignUpComplete: true,
       nextStep: AuthNextSignUpStep(signUpStep: 'DONE'),
     );
@@ -70,7 +60,7 @@ class AmplifyAuthCognitoStub extends AuthPluginInterface
     await Future<void>.delayed(const Duration(milliseconds: 50));
     _MockUser? user = _users[request.username];
     if (user == null) {
-      throw const UserNotFoundException('_MockUser does not exist.');
+      throw UserNotFoundException();
     }
     return CognitoResendSignUpCodeResult(_codeDeliveryDetails(user));
   }
@@ -80,7 +70,7 @@ class AmplifyAuthCognitoStub extends AuthPluginInterface
     await Future<void>.delayed(const Duration(milliseconds: 50));
     _MockUser? user = _users[request.username];
     if (user == null) {
-      throw const UserNotFoundException('_MockUser does not exist.');
+      throw UserNotFoundException();
     }
     if (user.password != request.password) {
       throw const NotAuthorizedException('Incorrect username or password.');
@@ -88,7 +78,7 @@ class AmplifyAuthCognitoStub extends AuthPluginInterface
     _currentUser = user;
     return CognitoSignInResult(
       isSignedIn: _isSignedIn(),
-      nextStep: AuthNextSignInStep(signInStep: 'DONE'),
+      nextStep: const AuthNextSignInStep(signInStep: 'DONE'),
     );
   }
 
@@ -97,7 +87,7 @@ class AmplifyAuthCognitoStub extends AuthPluginInterface
     await Future<void>.delayed(const Duration(milliseconds: 50));
     return CognitoSignInResult(
       isSignedIn: _isSignedIn(),
-      nextStep: AuthNextSignInStep(signInStep: 'DONE'),
+      nextStep: const AuthNextSignInStep(signInStep: 'DONE'),
     );
   }
 
@@ -124,7 +114,7 @@ class AmplifyAuthCognitoStub extends AuthPluginInterface
     }
     _MockUser? user = _users[request.username];
     if (user == null) {
-      throw const UserNotFoundException('_MockUser does not exist.');
+      throw UserNotFoundException();
     }
     return CognitoResetPasswordResult(
       isPasswordReset: true,
@@ -145,10 +135,10 @@ class AmplifyAuthCognitoStub extends AuthPluginInterface
     }
     _MockUser? user = _users[request.username];
     if (user == null) {
-      throw const UserNotFoundException('_MockUser does not exist.');
+      throw UserNotFoundException();
     }
     if (request.confirmationCode != '123456') {
-      throw const CodeMismatchException('Incorrect code. Please try again.');
+      throw CodeMismatchException();
     }
     _MockUser updatedUser = user.copyWith(password: request.newPassword);
     _users[request.username] = updatedUser;
@@ -289,19 +279,6 @@ class AmplifyAuthCognitoStub extends AuthPluginInterface
     );
   }
 }
-
-class _AuthStreamControllerStub extends AuthStreamController {
-  @override
-  StreamController<AuthHubEvent> get authStreamController {
-    return _authStreamController;
-  }
-}
-
-StreamController<AuthHubEvent> _authStreamController =
-    StreamController<AuthHubEvent>.broadcast(
-  onListen: () {},
-  onCancel: () {},
-);
 
 class _MockUser {
   final String sub;
